third_party/blink/web_tests/platform/generic/fast/xpath/invalid-resolver-expected.txt

PASS document.evaluate('/body', document, 5, 0, null) threw exception TypeError: Failed to execute 'evaluate' on 'Document': parameter 3 is not of type 'XPathNSResolver'..
PASS document.evaluate('/body', document, 'a resolver', 0, null) threw exception TypeError: Failed to execute 'evaluate' on 'Document': parameter 3 is not of type 'XPathNSResolver'..
PASS document.evaluate('/body', document, Date, 0, null).toString() is "[object XPathResult]"
PASS successfullyParsed is true

TEST COMPLETE
